feat: enhance Taskfile for improved development (#915)
* feat: enhance README and Taskfile for improved local development setup

This commit updates the README.md to include new sections on task usage for local development, detailing commands for spinning up single-node and multi-node environments. It also introduces a new "Taskfile.yml" structure, adding environment variables and tasks for managing local deployments, including commands for starting and stopping services like the kwild server, Postgres, and the indexer. These changes aim to streamline the development process and improve clarity for users setting up their local environments.

* fix: update Docker Compose and migration script for improved configuration and security

This commit modifies the `dev-indexer-compose.yaml` file to change the build context and Dockerfile path for the indexer service, enhancing the deployment structure. Additionally, it updates the `migrate.sh` script to include checks for required environment variables, ensuring that the migration process is secure and properly configured. These changes aim to improve the clarity and maintainability of the deployment setup.

* docs: update section title in README for clarity

* fix: update Taskfile.yml to use DEV_NODE_1_RPC_ENDPOINT for NODE_RPC_ENDPOINT

This commit modifies the `Taskfile.yml` to introduce the `DEV_NODE_1_RPC_ENDPOINT` variable, which is now used for the `NODE_RPC_ENDPOINT` in the environment configuration. This change enhances clarity and ensures that the correct RPC endpoint is utilized for local development tasks.

* docs: enhance Taskfile.yml with internal task descriptions

This commit adds descriptive comments to the Taskfile.yml, clarifying the purpose of the internal helper task `_compose` and the `devnet:env` task. Additionally, it provides context for the `devnet:gateway:start` task, improving the overall documentation and usability for developers working with the Taskfile.